snap: split metablock functionality up, fix reading of directories
