+++ snap/squashfs2/squashfs2.go
@@ -27,8 +27,9 @@ import (
 
 const (
 	// https://github.com/plougher/squashfs-tools/blob/master/squashfs-tools/squashfs_fs.h#L289
-	superblockSize    = 96
-	metadataBlockSize = 8192
+	superblockSize         = 96
+	metadataBlockSize      = 8192
+	directoryMaxEntryCount = 256
 
 	// Inode types supported by squashfs
 	inodeTypeDirectory         = 1
@@ -167,7 +168,6 @@ type metaBlockReader struct {
 type directory struct {
 	node    *squashfs_inode_dir
 	reader  *metaBlockReader
-	ref     metadataRef
 	loaded  bool
 	entries []squashfs_dir_entry
 }
@@ -364,7 +364,7 @@ func (sfs *SquashFileSystem) loadRootDirectory() error {
 	inode := &squashfs_inode_dir{}
 	inode.parse(inodeBuffer)
 
-	sfs.rootDirectory = sfs.directoryCreate(inode, sfs.superBlock.rootIno)
+	sfs.rootDirectory = sfs.directoryCreate(inode)
 	return nil
 }
 
@@ -377,11 +377,10 @@ func (sfs *SquashFileSystem) ReadFile(path string) ([]byte, error) {
 	return sfs.rootDirectory.readFile(path)
 }
 
-func (sfs *SquashFileSystem) directoryCreate(node *squashfs_inode_dir, ref metadataRef) *directory {
+func (sfs *SquashFileSystem) directoryCreate(node *squashfs_inode_dir) *directory {
 	return &directory{
 		node:   node,
 		reader: sfs.directoryReader,
-		ref:    ref,
 		loaded: false,
 	}
 }
@@ -416,23 +415,32 @@ func (d *directory) readDirectoryEntry() (squashfs_dir_entry, int) {
 }
 
 func (d *directory) loadEntries() error {
-	d.reader.seekToRef(d.ref)
+	d.reader.seek(int64(d.node.startBlock), int(d.node.offset))
 	println("loading directory entries", d.node.size)
 
-	bytesRead := 0
-	for bytesRead < int(d.node.size) {
-		squashfs_dir_header := d.readDirectoryHeader()
-		println("squashfs: directory header:", squashfs_dir_header.count, squashfs_dir_header.startBlock, squashfs_dir_header.ino)
+	if d.node.size == 3 {
+		// directory is empty
+		return nil
+	}
 
-		for i := 0; i < int(squashfs_dir_header.count); i++ {
+	bytesRead := 0
+	for bytesRead < int(d.node.size) - 3 {
+		dirHeader := d.readDirectoryHeader()
+		bytesRead += 12
+
+		println("squashfs: directory header:", dirHeader.count, dirHeader.startBlock, dirHeader.ino)
+		if dirHeader.count > directoryMaxEntryCount {
+			return fmt.Errorf("squashfs: invalid number of directory entries: %d", dirHeader.count)
+		}
+		
+		// squashfs is littered with magic arethmetics, count is
+		// actually one less than specified in count
+		for i := 0; i < int(dirHeader.count) + 1; i++ {
 			entry, size := d.readDirectoryEntry()
 			println("squashfs: directory entry:", entry.name)
 			d.entries = append(d.entries, entry)
 			bytesRead += size
-			break
 		}
-
-		bytesRead += 12
 	}
 
 	d.loaded = true
@@ -448,152 +456,12 @@ func (d *directory) readFile(path string) ([]byte, error) {
 		}
 	}
 
+	// find entry
+	for _, entry := d.entries {
+		entry
+	}
+
 	return nil, nil
-}
-
-// Metadata (inodes and directories) are compressed in 8Kbyte blocks.
-// Each compressed block is prefixed by a two byte length, the top bit is set if the block is uncompressed.
-// Inodes are packed into the metadata blocks, and are not aligned to block boundaries, therefore inodes
-// overlap compressed blocks. Inodes are identified by a 48-bit number which encodes the location of the
-// compressed metadata block containing the inode, and the byte offset into
-// that block where the inode is placed (<block, offset>).
-func metablockReaderCreate(stream *os.File, compression CompressionBackend, offset int64, ref ...metadataRef) *metaBlockReader {
-	m := &metaBlockReader{
-		stream:        stream,
-		streamOffset:  offset,
-		compression:   compression,
-		currentBlock:  0,
-		currentOffset: 0,
-	}
-
-	if len(ref) != 0 {
-		err := m.seekToRef(ref[0])
-		if err != nil {
-			return nil
-		}
-	}
-	return m
-}
-
-func (m *metaBlockReader) seekToRef(ref metadataRef) error {
-	if ref.offset < 0 || ref.offset >= metadataBlockSize {
-		return fmt.Errorf("squashfs: invalid metadata offset %d", ref.offset)
-	}
-	m.currentBlock = ref.block
-	m.currentOffset = ref.offset
-	return nil
-}
-
-type metablock struct {
-	position  int64
-	length    int
-	data      []byte
-	nextBlock int64
-}
-
-func min(a, b int) int {
-	if a < b {
-		return a
-	}
-	return b
-}
-
-func parseMetablockLength(data []byte) (int, bool) {
-	length := readUint16(data)
-	isCompressed := (length & 0x8000) == 0
-	length &= 0x7fff
-	return int(length), isCompressed
-}
-
-func (m *metaBlockReader) readMetablock(position int64) (*metablock, error) {
-	if _, err := m.stream.Seek(position, 0); err != nil {
-		return nil, err
-	}
-
-	// read the first two bytes, contains length of block and
-	// whether or not the data is compressed
-	buffer := make([]byte, 2)
-	if _, err := m.stream.Read(buffer); err != nil {
-		return nil, err
-	}
-
-	length, isCompressed := parseMetablockLength(buffer)
-	if length == 0 {
-		return nil, fmt.Errorf("squashfs: invalid metablock length")
-	}
-
-	// Store the length of the data into availableData which will be modified
-	// in case of compression. We want to keep length intact to calculate next
-	// block offset
-	availableData := length
-
-	// Read the data in any case, whether uncompressed or not. We determine
-	// afterwards what to do with it.
-	buffer = make([]byte, length)
-	bytesRead, err := m.stream.Read(buffer)
-	if err != nil {
-		return nil, err
-	}
-
-	if isCompressed {
-		if m.compression == nil {
-			return nil, fmt.Errorf("squashfs: no compression backend available")
-		}
-
-		// Decompressed data is always less or equal to 8KiB
-		decompressedBuffer := make([]byte, metadataBlockSize)
-		bytesDecompressed, err := m.compression.Decompress(buffer, decompressedBuffer)
-		if err != nil {
-			return nil, err
-		}
-
-		buffer = decompressedBuffer
-		availableData = bytesDecompressed
-	} else {
-		availableData = bytesRead
-	}
-
-	return &metablock{
-		position: position,
-		length:   availableData,
-		data:     buffer,
-
-		// The next block follows just after this one, which means
-		// the current position, the length header, and then all the data
-		// we read.
-		nextBlock: position + int64(length) + 2,
-	}, nil
-}
-
-func (m *metaBlockReader) read(buffer []byte) error {
-	block, err := m.readMetablock(m.streamOffset + m.currentBlock)
-	if err != nil {
-		return err
-	}
-
-	bytesRead := 0
-	for bytesRead < len(buffer) {
-		// If we have reached the end of the current block, read the next one
-		// It's important to note we can start several blocks behind target, so
-		// keep switching blocks until we reach the right one
-		for m.currentOffset >= block.length {
-			nextBlock, err := m.readMetablock(block.nextBlock)
-			if err != nil {
-				return err
-			}
-			m.currentBlock = block.position - m.streamOffset
-			m.currentOffset -= block.length
-			block = nextBlock
-		}
-
-		// Now that we have made sure we have enough data, copy it to the
-		// buffer and update the current offset
-		bytesToCopy := min(block.length-m.currentOffset, len(buffer)-bytesRead)
-		copy(buffer[bytesRead:], block.data[m.currentOffset:m.currentOffset+bytesToCopy])
-		m.currentOffset += bytesToCopy
-		bytesRead += bytesToCopy
-	}
-	return nil
 }
 
 func parseInode(data []byte) squashfs_inode {
